Completed all assigned tasks
Version 1.0.0 is ready

+++ src/main/java/com/alkl1m/taskmanager/controller/AuthController.java
@@ -1,9 +1,9 @@
 package com.alkl1m.taskmanager.controller;
 
 import com.alkl1m.taskmanager.dto.auth.*;
-import com.alkl1m.taskmanager.entity.RefreshToken;
-import com.alkl1m.taskmanager.entity.User;
+import com.alkl1m.taskmanager.entity.*;
 import com.alkl1m.taskmanager.enums.Role;
+import com.alkl1m.taskmanager.event.RegistrationCompleteEvent;
 import com.alkl1m.taskmanager.exception.TokenRefreshException;
 import com.alkl1m.taskmanager.repository.UserRepository;
 import com.alkl1m.taskmanager.service.auth.RefreshTokenService;
@@ -12,6 +12,8 @@ import com.alkl1m.taskmanager.util.JwtUtils;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
@@ -26,15 +28,18 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @AllArgsConstructor
 @RequestMapping("/api/auth")
+@Slf4j
 public class AuthController {
     private final AuthenticationManager authenticationManager;
     private final UserRepository userRepository;
     private final PasswordEncoder encoder;
     private final JwtUtils jwtUtils;
     private final RefreshTokenService refreshTokenService;
+    private final ApplicationEventPublisher publisher;
 
     @PostMapping("/login")
     public ResponseEntity<?> authenticateUser(@Valid @RequestBody LoginRequest loginRequest) {
+        log.info("Authenticating user: {}", loginRequest.email());
         Authentication authentication = authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(loginRequest.email(), loginRequest.password()));
         SecurityContextHolder.getContext().setAuthentication(authentication);
@@ -46,18 +51,16 @@ public class AuthController {
         RefreshToken refreshToken = refreshTokenService.createRefreshToken(userDetails.getId());
 
         ResponseCookie jwtRefreshCookie = jwtUtils.generateRefreshJwtCookie(refreshToken.getToken());
-
+        log.info("User {} authenticated successfully", loginRequest.email());
         return ResponseEntity.ok()
                 .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
                 .header(HttpHeaders.SET_COOKIE, jwtRefreshCookie.toString())
-                .body(new UserInfoResponse(userDetails.getId(),
-                        userDetails.getUsername(),
-                        userDetails.getEmail(),
-                        Role.USER));
+                .body(new MessageResponse("User logged in successfully!"));
     }
 
     @PostMapping("/signup")
-    public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signupRequest) {
+    public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signupRequest,
+                                          final HttpServletRequest request) {
         if (userRepository.existsByEmail(signupRequest.email())) {
             return ResponseEntity
                     .badRequest()
@@ -69,28 +72,29 @@ public class AuthController {
                     .badRequest()
                     .build();
         }
-
+        log.info("Registering user: {}", signupRequest.email());
         User user = new User(signupRequest.name(),
                 signupRequest.email(),
                 encoder.encode(signupRequest.password()),
                 Role.USER);
         userRepository.save(user);
-
+        publisher.publishEvent(new RegistrationCompleteEvent(user, applicationUrl(request)));
+        log.info("User {} registered successfully", signupRequest.email());
         return ResponseEntity.ok(new MessageResponse("User registered successfully!"));
     }
 
     @PostMapping("/refreshToken")
-    public ResponseEntity<?> refreshToken(@Valid @RequestBody HttpServletRequest request) {
+    public ResponseEntity<?> refreshToken(HttpServletRequest request) {
+        log.info("Refreshing token...");
         String refreshToken = jwtUtils.getJwtRefreshFromCookies(request);
-        System.out.println(refreshToken);
-
+        log.debug("Refresh token: {}", refreshToken);
         if ((refreshToken != null) && (refreshToken.length() > 0)) {
             return refreshTokenService.findByToken(refreshToken)
                     .map(refreshTokenService::verifyExpiration)
                     .map(RefreshToken::getUser)
                     .map(user -> {
                         ResponseCookie jwtCookie = jwtUtils.generateJwtCookie(user);
-
+                        log.info("Token refreshed successfully!");
                         return ResponseEntity.ok()
                                 .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
                                 .body(new MessageResponse("Token is refreshed successfully!"));
@@ -98,21 +102,33 @@ public class AuthController {
                     .orElseThrow(() -> new TokenRefreshException(refreshToken,
                             "Refresh token is not id database!"));
         }
+        log.warn("Refresh Token is empty!");
         return ResponseEntity.badRequest().body(new MessageResponse("Refresh Token is empty!"));
     }
 
     @PostMapping("/logout")
-    public ResponseEntity<?> logout() {
-        Object principle = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        Long userId = ((UserDetailsImpl) principle).getId();
-        refreshTokenService.deleteByUserId(userId);
-
+    public ResponseEntity<?> logout(HttpServletRequest request) {
+        log.info("Logging out user");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            Object principal = authentication.getPrincipal();
+            if (principal instanceof UserDetailsImpl) {
+                UserDetailsImpl userDetails = (UserDetailsImpl) principal;
+                Long userId = userDetails.getId();
+                refreshTokenService.deleteByUserId(userId);
+            }
+        }
         ResponseCookie jwtCookie = jwtUtils.getCleanJwtCookie();
         ResponseCookie jwtRefreshCookie = jwtUtils.getCleanJwtRefreshToken();
-
+        log.info("Cleaned JWT cookies. User signed out");
         return ResponseEntity.ok()
                 .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
                 .header(HttpHeaders.SET_COOKIE, jwtRefreshCookie.toString())
                 .body(new MessageResponse("You've been signed out!"));
+    }
+
+    public String applicationUrl(HttpServletRequest request){
+        return "http://" + request.getServerName() + ":" +
+                request.getServerPort() + request.getContextPath();
     }
 }
